requirements.txt
fastapi~=0.115.12
pydantic~=2.11.4
httpx~=0.28.1
uvicorn~=0.34.2

vertexai~=1.71.1
google-cloud-aiplatform~=1.71.1
protobuf~=5.29.4
pytest~=8.3.5
pytest-asyncio~=0.26.0
python-dotenv~=1.1.0